refactor: 타입 파일명 규칙 변경경
