+++ src/auth/dtos/jwt-access-token.dto.ts
@@ -1,5 +1,5 @@
 import { IsBoolean, IsDate, IsEnum, IsNumber, IsString } from 'class-validator';
-import { LoginPlatformType } from '../types/LoginPlatformType';
+import { LoginPlatformType } from '../types/login-platform.type';
 
 export class JwtAccessTokenDto {
   @IsNumber()
