+++ src/auth/dtos/jwt-refresh-token.dto.ts
@@ -1,5 +1,5 @@
 import { IsEnum, IsNumber } from 'class-validator';
-import { LoginPlatformType } from '../types/LoginPlatformType';
+import { LoginPlatformType } from '../types/login-platform.type';
 
 export class JwtRefreshTokenDto {
   @IsNumber()
