+++ src/auth/dtos/sns-account-user.dto.ts
@@ -1,5 +1,5 @@
 import { IsBoolean, IsEmail, IsEnum, IsString } from 'class-validator';
-import { LoginPlatformType } from '../types/LoginPlatformType';
+import { LoginPlatformType } from '../types/login-platform.type';
 
 export class snsAccountUserDto {
   @IsEnum(LoginPlatformType)
