+++ src/team/interfaces/team-stat.repository.interface.ts
@@ -1,11 +1,11 @@
-import { SortOrderType } from 'src/shared/types/SortOrderType';
-import { TeamStatsSortType } from '../types/TeamStatsSortType';
+import { SortOrderType } from 'src/shared/types/sort-order.type';
+import { TeamStatsType } from '../types/team-stats.type';
 import { TeamStat } from '../entities/team-stat.entity';
 
 export interface ITeamStatRepository {
   getTeamStats(
     year: number,
-    sortBy: TeamStatsSortType,
+    sortBy: TeamStatsType,
     sortOrder: SortOrderType,
     limit: number,
     offset: number,
